monir:修改侧滑窗的加载
